Work around xcb_translate_coordinates sign bug
xcb_translate_coordinates_reply_t has unsigned dst_{x,y}, so explicitly
cast them to int16_t to get the right values.
Reported on irc by frostwork.

xcb-proto was fixed in commit 661fe8dd.

Acked-by: Peter Hutterer <peter.hutterer@who-t.net>

+++ xwininfo.c
@@ -940,8 +940,8 @@ Display_Stats_Info (struct wininfo *w)
     if (!trans_coords)
 	Fatal_Error ("Can't get translated coordinates.");
 
-    rx = trans_coords->dst_x;
-    ry = trans_coords->dst_y;
+    rx = (int16_t)trans_coords->dst_x;
+    ry = (int16_t)trans_coords->dst_y;
     free (trans_coords);
 
     xright = (dw - rx - w->geometry->border_width * 2 -
